Move lock queries into protocol_level_locks module
